Add keybinding to select the group to switch my monitors to, by name
This is already set by the groups keymap, but it only allows you to
select the default 0-9 groups. If you want to have a completion menu
for finding groups by name, this keybinding will allow you to do that.

This kind of feature is one I use quite frequently in Emacs, as I
leave a server running, and having completion for a buffer that has
been sitting around for a long time is quite useful.