Add keybindings to manipulate windows, starting with renumbering them
This way I can renumber windows to be in the order I want them to be,
expecially when I open things in an order that I do not expect.

The stumpwm:renumber function takes in the number to replace the
currently focused window with, changing it. If the number to use as
the replacement is the same as another window's number, the two
numbers get swapped.
https://stumpwm.github.io/git/stumpwm-git_5.html