Define function to verify if a passed object is a boolean
Taken from https://alhassy.com/TypedLisp

The reason we want this is because I frequently interpolate a value to
a string, usually with the format procedure. Format will usually just
convert things without caring about what I passed in, meaning I could
pass an integer when I really want a string. For example,
```lisp
(defun example (cmd)
  (run-shell-command (format nil "~a" cmd)))
```

If I pass the number 5 to example, I would get an error. An error in
the window manager would hopefully cause a stack unwind at best, but
might log me out at worst.

Common Lisp allows asserting the type of a value in a procedure's
definition, which allows us to catch type errors earlier. Though, this
is probably not a particularly good definition of booleanp, but it
works on the few test cases we really care about.