Make mute toggling block
If the shell command is not synchronous/blocking, there is a race
condition between toggling the mute and fetching the current state of
the output device. Strangely, this would only happen when using the
speakers on my laptop, but not when using the 3.5mm auxiliary port for
my headphones.

Adding the trailing 't makes StumpWM run the shell command AND collect
its output, executing the command synchronously. This prevents the
race condition.