Create new StumpWM command type for port numbers
This lets us treat the input from the user as if it were a valid
port-number. This makes the code more robust, because if the user enters
either a non-sensical port number (x < 0 && x > 65535) or just a string,
we can handle it and print a message back out to the user informing them
of their mistake.